+++ src/main/java/com/vaadin/flow/component/grid/GridArrayUpdater.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License. 
  */
-package com.vaadin.flow.data.provider;
+package com.vaadin.flow.component.grid;
 
 import java.io.Serializable;
